pyproject.toml
[project]
name = "repl-nix-textchessserver"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "twilio>=9.3.8",
    "python-chess>=1.999",
]
